@@ -1,7 +1,11 @@
 import React from 'react';
-import ReactDom from 'react-dom';
-import App from 'App';
+import ReactDOM from 'react-dom';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
 
-import 'index.css';
+ReactDOM.render(<App />, document.getElementById('root'));
 
-ReactDom.render(<App />, document.getElementById('app'));
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: http://bit.ly/CRA-PWA
+serviceWorker.unregister();
